feet:Amem Jesus

+++ src/main/java/com/example/fast_stock/model/ItemNotaFiscalEntrada.java
@@ -29,6 +29,6 @@ public class ItemNotaFiscalEntrada {
     private Integer quantidade;
 
     @Column(nullable = false)
-    private BigDecimal precoUnitario;
+    private Long precoUnitario;
 
 }
